Remove hindent-mode
Since this introduces another external dependency on a system
binary (which pulls in the `.el` library), it constantly breaks on
systems without a Haskell ecosystem.

This is at odds with the purpose of this dotfile, which is to be able
to clone it on any system and have Emacs self-configure automatically.