Provide a mechanism to inhibit use-package's :ensure keyword when starting emacs
I use this on Nix systems in order to force me to add packages using
the Emacs derivation's argument for doing so. At the same time, I do want
my Emacs config to bootstrap itself using `:ensure` on non-Nix systems,
as it did previously.

The new mechanism relies on an environment variable
`EMACS_USE_PACKAGE_INHIBIT_ENSURE`, which, if set to `1`, disables :ensure
within Emacs. If set to 0, it explicitly forces `:ensure` to be
considered. The default behavior is influenced by whether Emacs is
installed via Nix or not: If the executable that started Emacs is in
the system's Nix store, we inhibit `:ensure` by default. Otherwise,
the default is to not inhibit.

Note that even if `:ensure` is not inhibited, each invocation of
`use-package` may be independently disabled by setting `:ensure nil` as
usual. My new mechanism just disables it in cases where `use-package` actually
would download a package.

Nix does provide an alternative method that parses a given `.emacs` or
org-tangled init file to autmatically discover uses of `use-package`.
I don't like this approach, because it couples the state of my
system's nixosConfiguration to my Emacs configuration, and I want to
tinker with them both separately. Installing just the packages through
Nix is a good middle ground, giving me e.g. predictable rollbacks if
packages break upstream, while ensuring that packages that rely on
specific binaries have their dependencies met.