sideline-blame: Replace blamer with sideline-blame
I did not like the way that blamer's overlay inserted new lines or
scrolled the page to the right on long lines.

sideline-blame instead picks a suitable empty line close to the point
as a target for showing the text in the overlay, and has no obtrusive
scrolling from what I have experienced.

Using sideline, we can also port flycheck, flymake and lsp-ui-sideline
to use it in future commits.