haskell-mode: Change lsp-wrapper server shell.nix lookup to use project.el
The existing code contained a typo which meant that it already did not
work with projectile.

This _ought_ to restore the behavior, but lacking a suitable Haskell
project with an old-style shell.nix file, I have yet to test it.

In the future, this could be extended to modern flake-style setups.

It's also not obvious whether I need this once I migrate from lsp-mode
to Eglot.