outline-magic: Remove outline-magic
This package was interfering with the behavior of `org-cycle` (TAB),
causing the setting of `org-cycle-emulate-tab` to be ignored. This had
the result that _all_ TAB keypresses folded the nearest heading, and
it was impossible to indent or use org tables.

I wan't using this anyway, and it seems to not have had an update
since 2018.